Win64: Fixup WORD_T and UL suffixes for llp64 data model
The #if defined(__MINGW64__)||defined(__MINGW32__) check
can probably be removed.  If that were to happen ensure
to keep the #if defined(__UINTPTR_TYPE__) part.

The __MINGW64__ guard is still included in this commit to
reduce impact to other users in lieu of a review and wider
testing.

The UINTPTR_C() and INTPTR_C() are defined to allow non-GCC
compilers to have a go at building in the same way they
previously did.

The result should be low impact for existing users and only
the MINGW cases being affected.

To see the pre-processor provided defines from GCC:
 echo '#include <stdint.h>' > file.c
 gcc -dM -E file.c

On Linux x86_64 (gcc 8.5.0) this looks like:
 #define __UINTPTR_TYPE__ long unsigned int
 #define __INTPTR_WIDTH__ 64
 #define UINTPTR_MAX (18446744073709551615UL)
 #define INTPTR_MAX (9223372036854775807L)
 #define INTPTR_MIN (-9223372036854775807L-1)
 #define __INTPTR_MAX__ 0x7fffffffffffffffL
 #define __INTPTR_TYPE__ long int
 #define __UINTPTR_MAX__ 0xffffffffffffffffUL

On Linux i686 (gcc 4.8.5) this looks like:
 #define UINTPTR_MAX (4294967295U)
 #define INTPTR_MAX (2147483647)
 #define INTPTR_MIN (-2147483647-1)
 #define __INTPTR_MAX__ 2147483647
 #define __INTPTR_TYPE__ int
 #define __UINTPTR_MAX__ 4294967295U
 #define __UINTPTR_TYPE__ unsigned int

On MINGW64 (gcc 12.2.0) this looks like:
 #define __MINGW32__ 1   /* yes this exists too */
 #define __MINGW64__ 1
 #define __UINTPTR_TYPE__ long long unsigned int
 #define __INTPTR_MAX__ 0x7fffffffffffffffLL
 #define INTPTR_MAX INT64_MAX
 #define INTPTR_MIN INT64_MIN
 #define _INTPTR_T_DEFINED
 #define __INTPTR_TYPE__ long long int
 #define __INTPTR_WIDTH__ 64
 #define __UINTPTR_MAX__ 0xffffffffffffffffULL
 #define UINTPTR_MAX UINT64_MAX
 #define _UINTPTR_T_DEFINED

On MINGW32 (gcc 12.2.0) this looks like:
 #define __MINGW32__ 1
 #define __UINTPTR_TYPE__ unsigned int
 #define __INTPTR_MAX__ 0x7fffffff
 #define INTPTR_MAX INT32_MAX
 #define INTPTR_MIN INT32_MIN
 #define _INTPTR_T_DEFINED
 #define __INTPTR_TYPE__ int
 #define __INTPTR_WIDTH__ 32
 #define __UINTPTR_MAX__ 0xffffffffU
 #define UINTPTR_MAX UINT32_MAX
 #define _UINTPTR_T_DEFINED